==> Projet/7_Bert_mixed_datasets.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
import time
import joblib
from sklearn.metrics import accuracy_score, classification_report, confusion_matrix

# Charger les embeddings des datasets
embeddings_fake_news_df = pd.read_csv("data/Embedded/embeddings_fake_news_base_bert.csv")
embeddings_isot_df = pd.read_csv("data/Embedded/embeddings_isot_base_bert.csv")
embeddings_fake_real_df = pd.read_csv("data/Embedded/embeddings_fake_real_base_bert.csv")

# "New data 1" : True de Fake-News et Fake de ISOT

On se propose de tester ce que cela donne lorsque l'on entraine sur un mix des datasets, comme nous l'avons fait pour le TF-IDF

# Création de notre dataset
embeddings_isot_df_sub = embeddings_isot_df.sample(frac=0.5, random_state=42)
true_from_fake_news = embeddings_fake_news_df[embeddings_fake_news_df['label'] == 0]
fake_from_isot = embeddings_isot_df_sub[embeddings_isot_df_sub['label'] == 1]

combined_df = pd.concat([true_from_fake_news, fake_from_isot], ignore_index=True)
print(combined_df['label'].value_counts())

# Réalisation du modèle
X = combined_df.drop(columns=['label']).values
y = combined_df['label'].values

# Split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

# Entraînement
classifier_combined = LogisticRegression(max_iter=1000)
start_time = time.time()
classifier_combined.fit(X_train, y_train)
print(f"Temps pour fit : {time.time() - start_time:.2f} secondes")

# Évaluation
y_pred = classifier_combined.predict(X_test)
print("Accuracy:", accuracy_score(y_test, y_pred))
print(classification_report(y_test, y_pred))
print("\nConfusion Matrix:\n", confusion_matrix(y_test, y_pred))

# On enregistre le modèle
joblib.dump(classifier_combined, "Modele/base_bert_logistic_combined.pkl")

# Test sur fake_real

X_fake_real = embeddings_fake_real_df.drop(columns=['label']).values
y_fake_real = embeddings_fake_real_df['label'].values

# Prédictions
y_fake_real_pred = classifier_combined.predict(X_fake_real)

# Évaluation
print("Accuracy:", accuracy_score(y_fake_real, y_fake_real_pred))
print(classification_report(y_fake_real, y_fake_real_pred))
print("\nConfusion Matrix:\n", confusion_matrix(y_fake_real, y_fake_real_pred))

Et ceci est étonamment mauvais...

# "New data 2" : Fake de Fake-News et True de ISOT

# Création de notre dataset
fake_from_fake_news = embeddings_fake_news_df[embeddings_fake_news_df['label'] == 1]
true_from_isot = embeddings_isot_df_sub[embeddings_isot_df_sub['label'] == 0]

combined_df = pd.concat([fake_from_fake_news, true_from_isot], ignore_index=True)
print(combined_df['label'].value_counts())

# Réalisation du modèle
X = combined_df.drop(columns=['label']).values
y = combined_df['label'].values

# Split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

# Entraînement
classifier_combined = LogisticRegression(max_iter=1000)
start_time = time.time()
classifier_combined.fit(X_train, y_train)
print(f"Temps pour fit : {time.time() - start_time:.2f} secondes")

# Évaluation
y_pred = classifier_combined.predict(X_test)
print("Accuracy:", accuracy_score(y_test, y_pred))
print(classification_report(y_test, y_pred))
print("\nConfusion Matrix:\n", confusion_matrix(y_test, y_pred))

# Test sur fake_real

X_fake_real = embeddings_fake_real_df.drop(columns=['label']).values
y_fake_real = embeddings_fake_real_df['label'].values

# Prédictions
y_fake_real_pred = classifier_combined.predict(X_fake_real)

# Évaluation
print("Accuracy:", accuracy_score(y_fake_real, y_fake_real_pred))
print(classification_report(y_fake_real, y_fake_real_pred))
print("\nConfusion Matrix:\n", confusion_matrix(y_fake_real, y_fake_real_pred))

# Mélange des deux datasets : "New data 3"

# Création de dataset mélangé

# Séparer les vraies et fausses news
isot_real = embeddings_isot_df_sub[embeddings_isot_df_sub['label'] == 1]
isot_fake = embeddings_isot_df_sub[embeddings_isot_df_sub['label'] == 0]

fake_news_real = embeddings_fake_news_df[embeddings_fake_news_df['label'] == 1]
fake_news_fake = embeddings_fake_news_df[embeddings_fake_news_df['label'] == 0]

# On sépare aléatoirement nos datasets
isot_real_1 = isot_real.sample(frac=0.5, random_state=42)
isot_real_2 = isot_real.drop(isot_real_1.index)

isot_fake_1 = isot_fake.sample(frac=0.5, random_state=42)
isot_fake_2 = isot_fake.drop(isot_fake_1.index)

fake_news_real_1 = fake_news_real.sample(frac=0.5, random_state=42)
fake_news_real_2 = fake_news_real.drop(fake_news_real_1.index)

fake_news_fake_1 = fake_news_fake.sample(frac=0.5, random_state=42)
fake_news_fake_2 = fake_news_fake.drop(fake_news_fake_1.index)

# Créer mixed_data_1 avec 50 % des vraies et fausses informations de chaque dataset
mixed_data_1 = pd.concat([isot_real_1, isot_fake_1, fake_news_real_1, fake_news_fake_1], ignore_index=True)

# Créer mixed_data_2 avec le reste des observations non présentes dans mixed_data_1
mixed_data_2 = pd.concat([isot_real_2, isot_fake_2, fake_news_real_2, fake_news_fake_2], ignore_index=True)

# Mélanger les datasets pour éviter toute structure dans les données
mixed_data_1 = mixed_data_1.sample(frac=1, random_state=5).reset_index(drop=True)
mixed_data_2 = mixed_data_2.sample(frac=1, random_state=6).reset_index(drop=True)

# Vérification
print("Mixed Data 1:")
print(mixed_data_1['label'].value_counts())

print("\nMixed Data 2:")
print(mixed_data_2['label'].value_counts())

# Notre modèle sur mixed_data_1

# On réalise le split des données

X = mixed_data_1.drop(columns=['label']).values
y = mixed_data_1['label'].values

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

#On applique une régression logistique
classifier_mixed_1 = LogisticRegression(max_iter=1000)
classifier_mixed_1.fit(X_train, y_train)

# Et on évalue notre petit modèle
y_pred = classifier_mixed_1.predict(X_test)

print("Accuracy:", accuracy_score(y_test, y_pred))
print("\nClassification Report:\n", classification_report(y_test, y_pred))
print("\nConfusion Matrix:\n", confusion_matrix(y_test, y_pred))

# Test sur Fake_real
# Prédictions
y_fake_real_pred = classifier_mixed_1.predict(X_fake_real)

# Évaluation
print("Accuracy:", accuracy_score(y_fake_real, y_fake_real_pred))
print(classification_report(y_fake_real, y_fake_real_pred))
print("\nConfusion Matrix:\n", confusion_matrix(y_fake_real, y_fake_real_pred))

# Notre modèle sur mixed_data_2

# On réalise le split des données

X = mixed_data_2.drop(columns=['label']).values
y = mixed_data_2['label'].values

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

#On applique une régression logistique
classifier_mixed_2 = LogisticRegression(max_iter=1000)
classifier_mixed_2.fit(X_train, y_train)

# Et on évalue notre petit modèle
y_pred = classifier_mixed_2.predict(X_test)

print("Accuracy:", accuracy_score(y_test, y_pred))
print("\nClassification Report:\n", classification_report(y_test, y_pred))
print("\nConfusion Matrix:\n", confusion_matrix(y_test, y_pred))

# Test sur Fake_real
# Prédictions
y_fake_real_pred = classifier_mixed_2.predict(X_fake_real)

# Évaluation
print("Accuracy:", accuracy_score(y_fake_real, y_fake_real_pred))
print(classification_report(y_fake_real, y_fake_real_pred))
print("\nConfusion Matrix:\n", confusion_matrix(y_fake_real, y_fake_real_pred))

Conclusion de cette exploration : 
- le mélange entre ISOT et fake_news n'apporte de réel gains que si l'on prends des vraies nouvelles et des fake news
- Evalué sur fake_real, on a gain limité, de 7 points.
